[Refactoring ⚙️] Post Info Component로 별도 분리
