+++ src/Pages/Post/components/PostDetail/Components/PostDetailInfo/Components/PostDetailTime/PostDetailTime.tsx
@@ -1,6 +1,6 @@
 import "./PostDetailTime.Style.css";
-import { convertDateFormat } from "../../../../../../Util/UtilFunctions";
-import { PostDetailTimeProps } from "../../PostDetail.Types";
+import { convertDateFormat } from "../../../../../../../../Util/UtilFunctions";
+import { PostDetailTimeProps } from "../../../../PostDetail.Types";
 import React from "react";
 
 const PostDetailTime = ({
