+++ src/Pages/Post/components/PostDetail/Components/PostDetailInfo/PostDetailContent.Styles.css
@@ -1,0 +1,12 @@
+.ptdetail__info {
+  display: flex;
+  align-items: center;
+}
+
+.ptdetail__info-divider {
+  height: 60%;
+
+  margin: 0 2rem;
+
+  border: 1px solid var(--color-semi-grey);
+}
